flask/requirements.txt
asn1crypto==1.5.1
black==23.7.0
blinker==1.6.2
click==8.1.6
exceptiongroup==1.1.2
Flask==2.3.2
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
pg8000==1.29.8
platformdirs==3.9.1
pluggy==1.2.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.19
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.3.6
